refactor(html): `event-provenance.ts` moves in with its only importer
`main/events.ts` is the whole of its audience: nothing else imports it, the
package's exports map does not name it, and it has no tests of its own to move.
The two types it defines reach the worker protocol, but they do so through
`main/events.ts`, which re-exports them.
